Update DataBaseManager.py
Added Information for connecting to the  test database hosted on the RDB server.

Commented on what each function was doing.

Updated the login function to send back all information from the account in a JSON format.

Updated the SQL queries to fall inline with the column names in the database.